// src/libs/types/member.ts
import { Types } from "mongoose";
import { MemberExperience, MemberSpecialty, MemberStatus, MemberType } from "../enums/member.enum";
import { Request } from "express";
import { Session } from "express-session";

export interface Member {
    _id: Types.ObjectId;
    memberType: MemberType;
    memberStatus: MemberStatus;
    memberNick: string;
    memberPhone: string;
    memberPassword?: string;
    memberAddress?: string;
    memberDesc?: string;
    memberImage?: string;
    memberPoints: number;
    memberSpecialty?: MemberSpecialty;
    memberExperience?: MemberExperience;
    createdAt: Date;
    updatedAt: Date;
}

export interface MemberInput {
    memberType?: MemberType;
    memberStatus?: MemberStatus;
    memberNick: string;
    memberPhone: string;
    memberPassword: string;
    memberAddress?: string;
    memberDesc?: string;
    memberImage?: string;
    memberPoints?: number;
}

export interface MemberUpdateInput {
    _id: Types.ObjectId;

    memberStatus?: MemberStatus;
    memberNick?: string;
    memberPhone?: string;
    memberPassword?: string;
    memberAddress?: string;
    memberDesc?: string;
    memberImage?: string;

}

export interface LoginInput {
    memberNick: string;
    memberPassword: string;
}


export interface AdminRequest extends Request {
    member: Member;
    session: Session & { member: Member }
    file: Express.Multer.File;
    files: Express.Multer.File[];
}